Movidas funcionalidades de botão adicionar e verDetalhes para classe abstrata pesquisaView
