+++ gerencia_de_estoque/src/view/ModosPesquisa.java
@@ -1,7 +1,6 @@
 package view;
 
 enum ModosPesquisa {
-    LISTAR_FILIAIS,
     LISTAR_ESTOQUE_GERAL,
     LISTAR_ESTOQUE_FILIAL,
 }
